requirements.txt
python-telegram-bot==21.4
# Pin any other libraries if added later.
